# openapi.yaml
openapi: 3.1.0
info:
  title: Chotu API
  version: 0.1.0
  description: "Chotu is an open-source, self-hostable vehicle fuel log. This is
    the M1 auth surface. Every error uses the shared Error schema: a stable
    `code`, a human `message`, and optional `details`."
  license:
    name: AGPL-3.0-only
    identifier: AGPL-3.0-only
paths:
...
  /reconcile:
    get:
      operationId: reconcile
      summary: Run reconciliation checks over the caller's own data
      description: Requires authentication (FR-17.1). Read-only. Returns every finding
        with a record id, check code, and a value-free message. A clean dataset
        returns an empty list (FR-17.5).
      responses:
        "200":
          description: The findings
          content:
            application/json:
              schema:
                type: object
                required:
                  - findings
                properties:
                  findings:
                    type: array
                    items:
                      $ref: "#/components/schemas/ReconcileFinding"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /admin/reconcile:
    get:
      operationId: adminReconcile
      summary: Run reconciliation deployment-wide
      description: Admin only (FR-17.2, INV-9). Each finding carries a record id, the
        owning user id, and a check code — no fuel-entry field values and no
        message.
      responses:
        "200":
          description: The findings
          content:
            application/json:
              schema:
                type: object
                required:
                  - findings
                properties:
                  findings:
                    type: array
                    items:
                      $ref: "#/components/schemas/AdminReconcileFinding"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Admin role required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /admin/invitations:
    post:
      operationId: adminCreateInvitation
      summary: Create a single-use invitation
      description: Admin only (FR-3.2). The plaintext invitation token is returned
        once and then only stored hashed. Default expiry is 7 days.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                invitedRole:
                  type: string
                  enum:
                    - user
                    - admin
                  default: user
                expiresInSeconds:
                  type: integer
                  minimum: 1
                  maximum: 2592000
              required:
                - email
              additionalProperties: false
      responses:
        "201":
          description: The invitation
          content:
            application/json:
              schema:
                type: object
                required:
                  - invitation
                  - invitationToken
                  - note
                properties:
                  invitation:
                    type: object
                    required:
                      - id
                      - email
                      - invitedRole
                      - expiresAt
                    properties:
                      id:
                        type: string
                      email:
                        type: string
                        format: email
                      invitedRole:
                        type: string
                        enum:
                          - user
                          - admin
                      expiresAt:
                        type: string
                        format: date-time
                  invitationToken:
                    type: string
                  note:
                    type: string
        "400":
          description: Malformed body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Admin role required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /admin/settings:
    get:
      operationId: adminGetSettings
      summary: Read the deployment settings
      description: Admin only (FR-9.1).
      responses:
        "200":
          description: The current settings
          content:
            application/json:
              schema:
                type: object
                required:
                  - settings
                properties:
                  settings:
                    $ref: "#/components/schemas/DeploymentSettings"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Admin role required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      operationId: adminUpdateSettings
      summary: Update the deployment settings
      description: Admin only (FR-9.1). `allowedAuthMethods` must stay non-empty;
        dropping `password` is refused while any user could be stranded by it
        (FR-9.2). `defaultCurrencyCode` is fixed to USD in M1.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                deploymentName:
                  type: string
                  minLength: 1
                  maxLength: 200
                registrationPolicy:
                  type: string
                  enum:
                    - invite_only
                    - open
                    - sso_auto
                allowedAuthMethods:
                  type: array
                  items:
                    type: string
                    enum:
                      - password
                      - oidc
                  minItems: 1
                defaultUnitSystem:
                  type: string
                  enum:
                    - imperial
                    - metric
                defaultTimeZone:
                  type: string
                  minLength: 1
                fuelVolumePrecision:
                  type: integer
                  minimum: 1
                  maximum: 3
                sessionTtlSeconds:
                  type: integer
                  minimum: 60
                apiTokenTtlSeconds:
                  type: integer
                  minimum: 60
                  nullable: true
              additionalProperties: false
      responses:
        "200":
          description: The updated settings
          content:
            application/json:
              schema:
                type: object
                required:
                  - settings
                properties:
                  settings:
                    $ref: "#/components/schemas/DeploymentSettings"
        "400":
          description: Malformed body or no field to change
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Admin role required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description: Removing password would leave a user with no sign-in method
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /admin/users:
    get:
      operationId: adminListUsers
      summary: List every account
      description: Admin only. Accounts, roles, status, and a vehicle count. No fuel
        entry data is ever included (INV-9).
      responses:
        "200":
          description: Every account, newest first
          content:
            application/json:
              schema:
                type: object
                required:
                  - users
                properties:
                  users:
                    type: array
                    items:
                      $ref: "#/components/schemas/AdminUserListItem"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Admin role required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      operationId: adminCreateUser
      summary: Create an account directly
      description: Admin only (FR-8.3). Omit `password` to get a one-time set-password
        link in the response instead.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                displayName:
                  type: string
                  minLength: 1
                  maxLength: 100
                role:
                  type: string
                  enum:
                    - user
                    - admin
                  default: user
                password:
                  type: string
                  minLength: 8
              required:
                - email
                - displayName
              additionalProperties: false
      responses:
        "201":
          description: The new account
          content:
            application/json:
              schema:
                type: object
                required:
                  - user
                properties:
                  user:
                    $ref: "#/components/schemas/AdminUserListItem"
                  setPasswordToken:
                    type: string
                    description: Present when no password was supplied
                  note:
                    type: string
        "400":
          description: Malformed body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Admin role required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /admin/users/{id}:
    get:
      operationId: adminGetUser
      summary: One account with security metadata
      description: Admin only. Adds last sign-in, linked identities, and the active
        API-token count. No fuel entry data is ever included (INV-9).
      parameters:
        - *a1
      responses:
        "200":
          description: The account
          content:
            application/json:
              schema:
                type: object
                required:
                  - user
                properties:
                  user:
                    $ref: "#/components/schemas/AdminUserDetail"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Admin role required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: No such user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: adminDeleteUser
      summary: Delete an account and all its data
      description: Admin only (FR-8.6). `confirmEmail` must equal the target's email.
        Cascades vehicles, entries, tokens, sessions, identities. Refused for
        the last active admin (INV-6).
      parameters:
        - *a1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                confirmEmail:
                  type: string
                  format: email
              required:
                - confirmEmail
              additionalProperties: false
      responses:
        "204":
          description: Deleted
        "400":
          description: Missing or mismatched confirmEmail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Admin role required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: No such user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description: This is the last active admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /admin/users/{id}/deactivate:
    post:
      operationId: adminDeactivateUser
      summary: Deactivate an account
      description: Admin only (FR-8.4). Cuts the user's sessions and API tokens
        immediately. Refused for the last active admin (INV-6).
      parameters:
        - *a1
      responses:
        "204":
          description: Done (idempotent)
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Admin role required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: No such user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description: This is the last active admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    cookieAuth:
      type: apiKey
      in: cookie
      name: chotu_session
  schemas:
    Error:
      type: object
      required:
        - code
        - message
      additionalProperties: false
      properties:
        code:
          type: string
          enum:
            - validation_error
            - unauthorized
            - forbidden
            - password_change_required
            - not_found
            - email_taken
            - invitation_consumed
            - conflict
            - odometer_decrease
            - last_admin
            - auth_method_required
            - provider_in_use
            - rate_limited
            - internal_error
        message:
          type: string
        details:
          description: Optional structured detail
...
    ReconcileFinding:
      type: object
      required:
        - recordType
        - recordId
        - vehicleId
        - checkCode
        - message
      properties:
        recordType:
          type: string
          enum:
            - fuel_entry
            - vehicle
        recordId:
          type: string
        vehicleId:
          type:
            - string
            - "null"
        checkCode:
          type: string
          enum:
            - duplicate
            - orphaned
            - odometer_tie
            - odometer_decrease
            - missing_field
            - out_of_range
        message:
          type: string
...
security:
  - bearerAuth: []
  - cookieAuth: []